LayoutTests/fast/regions/auto-height/two-auto-height-regions.html

<!doctype html>
<html>
    <head>
        <!--
        First 2 regions are auto-height. First para should enter in first one, second para
        should enter in the second one.
        -->
        <script src="../../../fast/js/resources/js-test-pre.js"></script>
        <style>
            #p1, #p2 { -webkit-flow-into: article; }
            #p1 { -webkit-region-break-after: always; }
            #regionA { -webkit-flow-from: article; width: 50px; height: auto; border: 1px solid red; }
            #regionB { -webkit-flow-from: article; width: 70px; height: auto; border: 1px solid green; }
            #regionC { -webkit-flow-from: article; width: 100px; height: 100px; border: 1px solid blue; }
        </style>
    </head>
    <body>
        <p id="p1">Text inside p1.</p>
        <p id="p2">Text inside p2.</p>
        <div id="regionA"></div>
        <div id="regionB"></div>
        <div id="regionC"></div>
        <script>
            if (window.layoutTestController)
                window.layoutTestController.dumpAsText();
            description("Test for regions auto height prototype. First 2 regions are auto height.");
            var namedFlow = document.webkitGetFlowByName("article");
            var p1RegionList = namedFlow.getRegionsByContentNode(document.getElementById("p1"));
            var p2RegionList = namedFlow.getRegionsByContentNode(document.getElementById("p2"));
            shouldEvaluateTo("p1RegionList.length", 1);
            shouldEvaluateTo("p2RegionList.length", 1);
            shouldBeEqualToString("p1RegionList.item(0).id", "regionA");
            shouldBeEqualToString("p2RegionList.item(0).id", "regionB");

            document.getElementById("p1").style.visibility = "hidden";
            document.getElementById("p2").style.visibility = "hidden";
            document.getElementById("regionA").style.visibility = "hidden";
            document.getElementById("regionB").style.visibility = "hidden";
            document.getElementById("regionC").style.visibility = "hidden";

        </script>
        <script src="../../../fast/js/resources/js-test-post.js"></script>
    </body>
</html>
